Moved state into its own file
